+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 
 ### Changed
-
+- Refactored job group filter to use magic numbers for outlier bounds.
 
 ### Fixed